requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
alembic
psycopg2-binary
asyncpg